ARQ-56 Add LifecycleMethodExecutor SPI
A Callback from Arquillian when the TestFramework should invoke the method. This is used for vetoing @BeforeClass, @AfterClass, @Before, @After and @Test style methods on the Client side, but to execute them in Container.

Add TestCase to verify the correct behavior for JUnit integration with @Before and @After.

Current only possible in JUnit, not supported by TestNG 5.12.1