ARQ-196 Add new IOC Arquillian core
In the process of supporting multiple Containers in the same run, a new arquillian event core was needed. New core is more CDI/AtInject style with activation/deactivation of Context pr Thread.

Currently supporting 5 Scopes, Application, Suite, Class, Test and Container.

@Inject is used to mark a field as a InjectionPoint.
@Observes is used to mark a method as a Observer of the event type.

Instance is a injectable reference to read a Object stored in the Contexts.

InstanceProducer extends Instance with the posibility to write to a Context. When injecting a InstanceProvider a Scope is needed to determine which context to place the object. When using InstanceProvider.set, a event of the object type set is fired. This gives the extension a option to not bind to specific lifecycle events like Before/AfterTest, but rather the availability of a Object.

Event is a reference that can be used to fire events of a type. The event firing is done sync on the same thread.

Example:

   public static class TestObserverOne
   {
      private boolean wasCalled = false;

      @Inject @SuiteScoped
      private InstanceProducer<Object> object;

      @Inject
      private Event<Integer> event;

      @Inject
      private Instance<ContainerContext> context;

      public void on(@Observes String name)
      {
         try
         {
            context.get().activate("");
            object.set(new Object());
            event.fire(new Integer(100));
         }
         finally
         {
            context.get().deactivate();
         }
         wasCalled = true;
      }
   }